#30 Move utils folder in client
